ciso: improved portability
Types and structures can have different size and alignment depending on target system.
The ciso.c utility was apparently designed to be compiled on Windows 32-bit.
Header and Indexes becomes different when compiled on Linux 64-bit,
resulting in format discrepancy.

Note : this code preserves the original idea of writing a structure into a file.
However, it requires a little-endian CPU.